refactor: rename controller.js to filesController.js
